Select Computer selection and disable the rest
This is just a test of states of buttons, when the player selects an
action, the computer will "also select" an action, and the states of
the selected buttons and the rest of them will change.

Now, I need to find a way to simplify the code to make it more readable
and easier to understand. Specially in the change of states between
active and inactive